requirements.txt
# EcoBot v2 — Dependencies

# Core framework
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# AI — OpenAI-compatible SDK (Gemini & OpenAI)
openai>=1.0.0

# Database — PostgreSQL
psycopg2-binary>=2.9.9

# Email — Resend
resend>=2.0.0

# PDF reports
reportlab==4.0.4

# Image processing
Pillow==10.0.1

# Production server (ASGI)
uvicorn[standard]>=0.30.0
asgiref>=3.7.0

# Development tools (optional)
pytest==7.4.2
black==23.9.1
flake8==6.0.0
